Avoid race-conditions when releasing blocked executions
If two processes try to release blocked executions for the same key, for example,
if the concurrency limit is 2, and 2 jobs finish at the same time, both would try
to release the first one, and one of them would fail to do that. To avoid that,
select the first one but lock the record, and have the SELECT ... FOR UPDATE
to use SKIP LOCKED so we don't have to wait on the lock in other processes.