Simplify ready execution claiming, passing only job_ids
We no longer need to pass executions, as that was a change to
prepare for an implementation of sequential jobs that I won't
be pursuing. We can simply pluck the job_ids of the selected
executions, which as a side-effect, avoids an extra query due
to counting the jobs before actually trying to claim them.

Thanks to @djmb for spotting the extra `exists?` query that
was a consequence of this change.