Delegate concurrency limit and duration to job class and store `expires_at`
We don't need to store the concurrency limit and max duration as the
Solid Queue job has the job class name available to constantize it and
get them from there.

Also, stop unblocking executions in the worker before polling, keep that to
polling only. We'll have the supervisor do this, as well as expiring semaphores.