Add a Puma plugin for Solid Queue
This allows you to run Solid Queue as part of a puma cluster, by adding

```
plugin :solid_queue
```

to your Puma config.

It will boot up the Solid Queue supervisor, leaving it to manage the
scheduler and worker processes as normal.

If either Puma or the Solid Cache supervisor dies, the other will notice
and shut itself down.

On puma shutdown, it will signal Solid Queue to also shutdown.