Bring back support for prefix matching in queues assigned to a worker
This was removed in https://github.com/basecamp/solid_queue/pull/45, as the
resulting queries for polling using `queue_name LIKE 'something%'` were very
slow as they prevented MySQL from using an index for ordering. The idea after
removing that was to support namespaces so we could combine a namespace with
a "*" for queue names, to have support for a single prefix. However, namespaces
felt a too big of a change just to accomodate for our staging and beta setups,
so I've brought this back, but this time with a different implementation,
based on fetching the list of queue names relevant for polling, filtering there
with the prefix. This query is very fast because it can use an existing index
starting with queue_name and apply the Loose Index Scan technique to examine
as many rows as distinct values there are.