Fix race condition when releasing a blocked execution
We need to lock the blocked execution we've picked up as candidate
to be released for the whole transaction while releasing it. Otherwise,
we could select it and another job select it right after us, then both
would try to release it, leading to a duplicate key exception when trying
to promote to ready.