Rename Scheduler → Dispatcher
It's a much more accurate name, since this process doesn't schedule jobs or executions, these
are already scheduled. What it does is to dispatch them, move them to ready so the workers
can process them. This rename will be useful later when we have a dispatcher for the cron-style
jobs.