Don't go through the general dispatch flow when releasing claimed executions
That's it, don't try to gain the concurrency lock, because claimed executions with
concurrency limits that are released would most likely be holding the semaphore
themselves, as it's released after completing. This means these claimed executions
would go to blocked upon release, leaving the semaphore busy. Just assume that if
a job has a claimed execution, it's because it already gained the lock when going
to ready.