Corrects a minor bug in ScheduledExecution
ScheduledExecution#dispatch_next_batch was returning either an empty
array or an integer count of the number of rows processed by the method.
The client of this method is Worker and Dispatcher #poll, and poll
requires the count of the number of rows processed.

This commit changes the returned empty array to an integer value of 0
for consistency’s sake.  It also includes small updates to the tests
for these methods.