Fixes issue #482: Work with < Rubies 3.2
The current Thread::Queue based Interruptible can not work with
Ruby version earlier than 3.2.  Given that SQ sets it's minimum
supported version of Ruby is derived from "full support of Rails 7.1",
this in theory mandates support for Ruby 2.7.8.  However, other
dependencies force the minimum version of Ruby to: 3.1.6

This commit adds a boot time check of the Ruby version and selects
either the current or original implementation of Interruptible.